requirements.txt
tensorflow==2.12.0
numpy==1.24.2
Pillow==9.5.0
tqdm==4.64.1
scikit-learn==1.0.2
opencv-python==4.6.0.66
matplotlib==3.7.0
torch==2.0.0
efficientnet-pytorch==0.7.1
Pillow==8.0.0
pip uninstall torch
pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
